analyzer/requirements.txt
numpy==1.26.4
torch==2.2.2
transformers==4.38.2
python-dotenv==1.0.1
requests==2.31.0
kafka-python==2.1.2